+++ schemes/dc/rlwe_dc/src/PiDCRLWE.cpp
@@ -238,7 +238,7 @@ bool PVHSS_ACC_TEST(int msg_num, int degree_f)
     ZZ y_native = to_ZZ(MPE(X, degree_f) % param.ck.g1_order_ZZ);
     cout << "True result: " << y_native << endl;
     cout << "Eval result: " << y_eval << endl;
-    const ZZ decode_limit = conv<ZZ>(PVHSS_M_MAX);
+    const ZZ decode_limit = conv<ZZ>((unsigned long)PVHSS_M_MAX);
     if (y_native >= decode_limit)
     {
         cout << "P5 accuracy check relaxed: native result is outside DecPed 32-bit decryption range." << endl;
